fix retrieving openssl 3 versions
OpenSSL 3 now generates `include/openssl/opensslv.h` and that file
in Node.js now includes the C++ preprocessor macros that are parsed
to determine the OpenSSL version via two levels of indirection. Add
a fallback (we need to check the original file location first for
older release lines using pre-OpenSSL 3) to one of the generated
files in the `config` dir.